requirements.txt
Flask==2.3.3
PyYAML==6.0.1
kubernetes==28.1.0
scikit-learn==1.3.2
pandas==2.1.2
gunicorn==21.2.0
prometheus-client==0.19.0
requests==2.32.0
